Abstracted tower code to parent class for ease of use
